Refactor display_export to static methods
