+++ libraries/classes/DisplayResults.php
@@ -5101,7 +5101,7 @@ class DisplayResults
 
         // Export link
         // (the url_query has extra parameters that won't be used to export)
-        // (the single_table parameter is used in PMA_getExportDisplay()
+        // (the single_table parameter is used in Export::getExportDisplay()
         //  to hide the SQL and the structure export dialogs)
         // If the parser found a PROCEDURE clause
         // (most probably PROCEDURE ANALYSE()) it makes no sense to
