+++ server_export.php
@@ -6,6 +6,7 @@
  * @package PhpMyAdmin
  */
 use PhpMyAdmin\Config\PageSettings;
+use PhpMyAdmin\Display\Export;
 use PhpMyAdmin\Response;
 
 /**
@@ -15,7 +16,6 @@ require_once 'libraries/common.inc.php';
 require_once 'libraries/config/user_preferences.forms.php';
 require_once 'libraries/config/page_settings.forms.php';
 require_once 'libraries/server_common.inc.php';
-require_once 'libraries/display_export.lib.php';
 
 PageSettings::showGroup('Export');
 
@@ -27,9 +27,7 @@ $scripts->addFile('export.js');
 $export_page_title = __('View dump (schema) of databases') . "\n";
 
 $select_item = isset($tmp_select)? $tmp_select : '';
-$multi_values  = PMA_getHtmlForExportSelectOptions($select_item);
-
-require_once 'libraries/display_export.lib.php';
+$multi_values  = Export::getHtmlForExportSelectOptions($select_item);
 
 if (! isset($sql_query)) {
     $sql_query = '';
@@ -42,7 +40,7 @@ if (! isset($unlim_num_rows)) {
 }
 $response = Response::getInstance();
 $response->addHTML(
-    PMA_getExportDisplay(
+    Export::getExportDisplay(
         'server', $db, $table, $sql_query, $num_tables,
         $unlim_num_rows, $multi_values
     )
